refactor: centralize Scope definition in models.auth, update imports and references across the codebase
