+++ src/routes/quest_progress.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter, status, Security, Depends
 
-from src.dependencies.auth import Scope, get_guild_client
+from src.dependencies.auth import get_guild_client
 from src.dependencies.services import get_quest_progress_service
-from src.models.auth import TokenPayload
+from src.models.auth import TokenPayload, Scope
 
 from src.models.quests.quest_progress import QuestProgressIn, QuestProgressOut, QuestProgressUpdate
 from src.services.quest_progress import QuestProgressService
@@ -13,7 +13,7 @@ quest_progress_router = APIRouter(prefix='/guilds/me/quests_router/progress', ta
 @quest_progress_router.post('', status_code=status.HTTP_201_CREATED)
 async def create_quest_progress(
         body: QuestProgressIn,
-        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_WRITE]),
+        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_QUESTS_WRITE]),
         service: QuestProgressService = Depends(get_quest_progress_service)
 ) -> QuestProgressOut:
     """
@@ -28,7 +28,7 @@ async def create_quest_progress(
 @quest_progress_router.get('/user/{thorny_id}')
 async def list_quest_progress(
         thorny_id: int,
-        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_READ]),
+        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_QUESTS_READ]),
         service: QuestProgressService = Depends(get_quest_progress_service)
 ) -> list[QuestProgressOut]:
     """
@@ -42,7 +42,7 @@ async def list_quest_progress(
 @quest_progress_router.get('/user/{thorny_id}/active')
 async def get_active_quest_progress(
         thorny_id: int,
-        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_READ]),
+        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_QUESTS_READ]),
         service: QuestProgressService = Depends(get_quest_progress_service)
 ) -> QuestProgressOut:
     """
@@ -56,7 +56,7 @@ async def get_active_quest_progress(
 @quest_progress_router.delete('/user/{thorny_id}/active', status_code=status.HTTP_204_NO_CONTENT)
 async def fail_active_quest_progress(
         thorny_id: int,
-        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_WRITE]),
+        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_QUESTS_WRITE]),
         service: QuestProgressService = Depends(get_quest_progress_service)
 ):
     """
@@ -72,7 +72,7 @@ async def fail_active_quest_progress(
 async def partial_update_quest_progress(
         progress_id: int,
         body: QuestProgressUpdate,
-        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_WRITE]),
+        _: TokenPayload = Security(get_guild_client, scopes=[Scope.GUILDS_QUESTS_WRITE]),
         service: QuestProgressService = Depends(get_quest_progress_service)
 ) -> QuestProgressOut:
     """
